+++ mountpoint-s3-fs/tests/common/mod.rs
@@ -26,12 +26,12 @@ use mountpoint_s3_client::ObjectClient;
 use mountpoint_s3_fs::fs::{DirectoryEntry, DirectoryReplier};
 use mountpoint_s3_fs::prefetch::{default_prefetch, DefaultPrefetcher};
 use mountpoint_s3_fs::prefix::Prefix;
-use mountpoint_s3_fs::{S3Filesystem, S3FilesystemConfig};
+use mountpoint_s3_fs::{Runtime, S3Filesystem, S3FilesystemConfig};
 use std::collections::VecDeque;
 use std::future::Future;
 use std::sync::Arc;
 
-pub type TestS3Filesystem<Client> = S3Filesystem<Client, DefaultPrefetcher<ThreadPool>>;
+pub type TestS3Filesystem<Client> = S3Filesystem<Client, DefaultPrefetcher>;
 
 pub fn make_test_filesystem(
     bucket: &str,
@@ -60,7 +60,7 @@ pub fn make_test_filesystem_with_client<Client>(
 where
     Client: ObjectClient + Clone + Send + Sync + 'static,
 {
-    let runtime = ThreadPool::builder().pool_size(1).create().unwrap();
+    let runtime = Runtime::new(ThreadPool::builder().pool_size(1).create().unwrap());
     let prefetcher = default_prefetch(runtime.clone(), Default::default());
     S3Filesystem::new(client, prefetcher, runtime, bucket, prefix, config)
 }